refactor(scheduler): split implementation and share internals
