+++ packages/internal/test/vitest.config.mts
@@ -1,0 +1,7 @@
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+	test: {
+		include: ['./test/*.test.mts'],
+	},
+});
