+++ packages/types/package.json
@@ -1,0 +1,24 @@
+{
+	"name": "@slipher/types",
+	"version": "0.0.0",
+	"private": true,
+	"license": "MIT",
+	"files": [
+		"lib/**"
+	],
+	"types": "./src/index.ts",
+	"scripts": {
+		"dev": "tsc --watch",
+		"build": "tsc",
+		"lint": "biome lint --write ./src",
+		"format": "biome format --write ./src",
+		"checkb": "biome check --write ./src",
+		"typecheck": "tsc --noEmit",
+		"test": "pnpm typecheck",
+		"prepublishOnly": "exit 1"
+	},
+	"devDependencies": {
+		"@biomejs/biome": "2.3.11",
+		"typescript": "^5.9.3"
+	}
+}
