+++ packages/types/tsconfig.json
@@ -1,0 +1,15 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"emitDeclarationOnly": true,
+		"rootDir": "./src",
+		"outDir": "./lib"
+	},
+	"include": [
+		"src"
+	],
+	"exclude": [
+		"**/lib",
+		"**/test"
+	]
+}
